Refactor parsing pipeline
